MM: aggiunto algoritmo sensirion

+++ Inc/AppConst.h
@@ -4,7 +4,7 @@
 typedef struct {
     float temperature;
     float humidity;
-    uint16_t voc;
+    uint32_t voc;
 } SensorData_t;
 
 #endif
